fix(db): 트리플 추출 배치 실행 전 memory_item 컬럼 보정
- ensureMemoryItemTripleExtractionColumns를 전용 모듈로 분리
- TripleExtractionBatchJob.execute 시작 시 호출해 init/dist 불일치 시에도 DDL 적용

Made-with: Cursor
